requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.28
mysqlclient==2.1.1
protobuf==4.22.3
Werkzeug==2.2.3
